feat: ClosedQuestionBar, OpenQuestionBar컴포넌트 통합
